+++ src/functions-src/Entities/IDomain.ts
