+++ src/functions-src/Entities/Photo/AlbumId.ts
@@ -1,6 +1,6 @@
-import {ValueObject} from "../ValueObject";
+import {IValueObject} from "../IValueObject";
 
-export class AlbumId implements ValueObject<string | null> {
+export class AlbumId implements IValueObject<string | null> {
     private readonly _value: string | null;
 
     constructor(value: string | null) {
