+++ src/functions-src/Entities/Photo/Url.ts
@@ -1,6 +1,6 @@
-import {ValueObject} from "../ValueObject";
+import {IValueObject} from "../IValueObject";
 
-export class Url implements ValueObject<string | null> {
+export class Url implements IValueObject<string | null> {
     private readonly _value: string | null;
 
     constructor(value: string | null) {
