+++ src/functions-src/Entities/Token/RefreshToken.ts
@@ -1,14 +1,14 @@
-import {ValueObject} from "../ValueObject";
+import {IValueObject} from "../IValueObject";
 
-export class AccessToken implements ValueObject<string | null> {
+export class RefreshToken implements IValueObject<string | null> {
     private readonly _value: string | null;
 
     constructor(value: string | null) {
         this._value = value;
     }
 
-    public static of(value: string | null): AccessToken {
-        return new AccessToken(value);
+    public static of(value: string | null): RefreshToken {
+        return new RefreshToken(value);
     }
 
     public value(): string | '' {
