+++ src/functions-src/Entities/Track/Artist.ts
@@ -1,6 +1,6 @@
-import {ValueObject} from "../ValueObject";
+import {IValueObject} from "../IValueObject";
 
-export class Artist implements ValueObject<string | null> {
+export class Artist implements IValueObject<string | null> {
     private readonly _value: string | null;
 
     constructor(value: string | null) {
