+++ src/functions-src/Entities/Track/Link.ts
@@ -1,14 +1,14 @@
-import {ValueObject} from "../ValueObject";
+import {IValueObject} from "../IValueObject";
 
-export class Name implements ValueObject<string | null> {
+export class Link implements IValueObject<string | null> {
     private readonly _value: string | null;
 
     constructor(value: string | null) {
         this._value = value;
     }
 
-    public static of(value: string | null): Name {
-        return new Name(value);
+    public static of(value: string | null): Link {
+        return new Link(value);
     }
 
     public value(): string | '' {
