+++ src/functions-src/Repositories/PhotoRepository/IPhotoRepository.ts
@@ -1,6 +1,6 @@
-import {Photo} from "../../Domains/Photo/Photo";
-import {AlbumId} from "../../Domains/Photo/AlbumId";
-import {AccessToken} from "../../Domains/Token/AccessToken";
+import {Photo} from "../../Entities/Photo/Photo";
+import {AlbumId} from "../../Entities/Photo/AlbumId";
+import {AccessToken} from "../../Entities/Token/AccessToken";
 
 export interface IPhotoRepository {
     // queries
