+++ src/functions-src/Services/Photo/GooglePhotoService.ts
@@ -1,6 +1,6 @@
-import {AlbumId} from "../../Domains/Photo/AlbumId";
-import {Photo} from "../../Domains/Photo/Photo";
-import {Url} from "../../Domains/Photo/Url";
+import {AlbumId} from "../../Entities/Photo/AlbumId";
+import {Photo} from "../../Entities/Photo/Photo";
+import {Url} from "../../Entities/Photo/Url";
 import {IPhotoService} from "./IPhotoService";
 import {IPhotoRepository} from "../../Repositories/PhotoRepository/IPhotoRepository";
 import {ITokenService} from "../Token/ITokenService";
