+++ src/functions-src/Services/Photo/IPhotoService.ts
@@ -1,5 +1,5 @@
-import {AlbumId} from "../../Domains/Photo/AlbumId";
-import {Photo} from "../../Domains/Photo/Photo";
+import {AlbumId} from "../../Entities/Photo/AlbumId";
+import {Photo} from "../../Entities/Photo/Photo";
 
 export interface IPhotoService {
     getARandomPhotoFromAlbum(albumId: AlbumId): Promise<Photo>;
