+++ src/functions-src/Services/Token/ITokenService.ts
@@ -1,5 +1,5 @@
-import {Token} from "../../Domains/Token/Token";
-import {AccessToken} from "../../Domains/Token/AccessToken";
+import {Token} from "../../Entities/Token/Token";
+import {AccessToken} from "../../Entities/Token/AccessToken";
 
 export interface ITokenService {
     getAccessAndRefreshToken(): Promise<Token>;
