+++ src/functions-src/Services/Token/TokenService.ts
@@ -1,8 +1,8 @@
 import {AxiosStatic} from "axios";
 import {ITokenService} from "./ITokenService";
-import {Token} from "../../Domains/Token/Token";
-import {AccessToken} from "../../Domains/Token/AccessToken";
-import {RefreshToken} from "../../Domains/Token/RefreshToken";
+import {Token} from "../../Entities/Token/Token";
+import {AccessToken} from "../../Entities/Token/AccessToken";
+import {RefreshToken} from "../../Entities/Token/RefreshToken";
 import {IOAuthConfig, ITokenRepository} from "../../Repositories/TokenRepository/ITokenRepository";
 
 export class TokenService implements ITokenService {
