+++ src/functions-src/Services/Track/ITrackService.ts
@@ -1,4 +1,4 @@
-import {Track} from "../../Domains/Track/Track";
+import {Track} from "../../Entities/Track/Track";
 
 export interface ITrackService {
     getCurrentlyListeningTrack(): Promise<Track>;
