+++ src/functions-src/Services/Track/SpotifyService.ts
@@ -1,4 +1,4 @@
-import {Track} from "../../Domains/Track/Track";
+import {Track} from "../../Entities/Track/Track";
 import {ITrackService} from "./ITrackService";
 import {ITrackRepository} from "../../Repositories/TrackRepository/ITtrackRepository";
 import {ITokenService} from "../Token/ITokenService";
